pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "radbot"
version = "0.1.0"
authors = [
    {name = "Perry Manuk", email = "perry.manuk@example.com"},
]
description = "A modular AI agent framework using Google ADK, Qdrant, MCP, and A2A"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "google-adk>=0.4.0",
    "google-genai>=1.11.0",  # New recommended Google GenAI SDK
    "qdrant-client==1.7.0",  # Pinned to a version that works well with our implementation
    "modelcontextprotocol>=0.1.0",
    "zoneinfo; python_version < '3.9'",
    "tzdata",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "nest-asyncio>=1.6.0",  # Required for MCP fileserver to work in running event loops
    "requests>=2.31.0",     # Required for Crawl4AI API calls
    "pymilvus>=2.4.0",      # Milvus client for vector database
    "sentence-transformers>=2.5.0",  # For generating embeddings
    "websockets>=12.0.0",   # Required for Home Assistant WebSocket client
    "httpx>=0.27.0",        # HTTP client for async requests
    "crawl4ai<=0.6.2",      # Required for proper deep crawling (using latest available version)
    "beautifulsoup4>=4.12.0", # Required for HTML parsing during crawling
    "aiohttp>=3.9.0",       # Required for async HTTP requests in crawling
    "psycopg2-binary>=2.9.9",  # Required for PostgreSQL database connections
    "redis>=5.0.1",         # Optional for cross-session caching with Redis
    "google-api-python-client>=2.102.0", # Required for Google Calendar API
    "google-auth-httplib2>=0.1.0",      # Required for Google Calendar API authentication
    "google-auth-oauthlib>=1.1.0",      # Required for Google Calendar OAuth flow
    "langchain-community>=0.2.16",      # Required for integrations
    "tavily-python>=0.3.8",             # Required for Tavily search integration
    "python-jsonrpc>=0.10.0",          # Required for Context7 MCP stdio client
    # Removed gcsa due to dependency conflict with ADK 0.4.0 (tzlocal version conflict)
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.5.0",
    "pytest>=8.3.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
]

web = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "jinja2>=3.1.4",
    "starlette>=0.36.0",
    "python-multipart>=0.0.9",
]

[project.urls]
"Homepage" = "https://github.com/perrymanuk/radbot"
"Bug Tracker" = "https://github.com/perrymanuk/radbot/issues"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]